Introduce a new MessageBuilder class for each message.
MessageBuilder is a mutable class providing the ability to merge a
message from a CodedInputStream into the mutable state, and build an
instance of a message.

MessageCompanion.newBuilder returns a new Builder initialized with empty
values.

MessageCompanion.newBuilder(msg) returns a new Builder initialized
with values from a given instance of a message.

This new functionality allows us to create builders with partially
initialized objects that can not be represented by the actual message.

See scalapb/ScalaPB#1013 and scalapb/scalapb-validate#38